front-end-projects/css-animations/requirements.txt
pywebview>=4.0.0



